+++ compare-frontend/src/app/modules/wizard/tabComponents/error/error.component.ts
@@ -1,7 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
-import { ScenarioName, Partition } from 'src/app/state/state';
+import { ScenarioName, Partition, RiesgosState } from 'src/app/state/state';
 import { WizardComposite } from '../../wizard.service';
 import { StepStateError } from 'src/app/state/state';
+import { Store } from '@ngrx/store';
+import { stepReset } from 'src/app/state/actions';
 
 @Component({
   selector: 'app-error',
@@ -12,12 +14,24 @@ export class ErrorComponent implements OnInit {
   @Input() scenario!: ScenarioName;
   @Input() partition!: Partition;
   @Input() data!: WizardComposite;
-  public message: any = {};
+  public message = "";
 
+  constructor(private store: Store<RiesgosState>) {}
 
   ngOnInit(): void {
     if (this.data.step.state.type === 'error') {
-      this.message = JSON.parse((this.data.step.state as StepStateError).message);
+      let message = (this.data.step.state as StepStateError).message;
+      if (typeof message === "string") {
+        message.replace("\n", "<br/>");
+      }
+      if (typeof message !== "string") {
+        message = JSON.parse(message);
+      }
+      this.message = message;
     }
+  }
+
+  public retry() {
+    this.store.dispatch(stepReset({scenario: this.scenario, partition: this.partition, stepId: this.data.step.step.id }));
   }
 }