+++ compare-frontend/src/app/state/effects.ts
@@ -1,9 +1,8 @@
 import { Injectable } from "@angular/core";
 import { Actions, createEffect, ofType } from "@ngrx/effects";
 import { Store } from "@ngrx/store";
-import { forkJoin, of } from "rxjs";
-import { catchError, delay, filter, map, mergeMap, switchMap, tap, withLatestFrom } from "rxjs/operators";
-import { BackendService } from "../services/backend.service";
+import { delay, filter, map, mergeMap, switchMap, tap, withLatestFrom } from "rxjs/operators";
+import { BackendService, isExecError } from "../services/backend.service";
 import { ConfigService } from "../services/config.service";
 import { ResolverService } from "../services/resolver.service";
 import * as AppActions from "./actions";
@@ -57,37 +56,31 @@ export class Effects {
             // must be *merge*-map for multiple requests in parallel to not interfer with one another
             mergeMap(({action, apiState}) => {                        
                 return this.backendSvc.execute(action.scenario, action.step, apiState).pipe(
-
-                    // instrumenting results with the action that started the execution
-                    map(apiScenarioState => ({ action, apiScenarioState })),
-
-                    tap(({action}) => console.log(`Execute success: ${action.scenario}/${action.partition}/${action.step}`)),
-
-                    // instrumenting potential errors with the action that started the execution
-                    catchError((err, caught) => {
-                        throw new ExecutionError(err, action.scenario, action.partition, action.step);
+                    map(result => {
+                        if (isExecError(result)) {
+                            // instrumenting results with the action that started the execution
+                            return new ExecutionError(JSON.stringify(result.error), action.scenario, action.partition, action.step);
+                        }
+                        console.log(`Execute success: ${action.scenario}/${action.partition}/${action.step}`);
+                        return {action, state: result};
                     })
-
-                    // NOTE: instrumenting this pipe with an action that is only used far down the line seems inelegant. 
-                    // Maybe add the action to a zone instead?
                 );
             })
         );
 
-        const convertedResults$ = executeResults$.pipe(map(({action, apiScenarioState}) => ({
-            newData: convertApiDataToRiesgosData(apiScenarioState.data),
-            action: action
-        })));
-
-        const successAction$ = convertedResults$.pipe(map(({newData, action}) => AppActions.stepExecSuccess({ scenario: action.scenario, partition: action.partition, step: action.step, newData })));
-
-        const caughtAction$ = successAction$.pipe(
-            catchError((err: ExecutionError, caught) => {
-                return of(AppActions.stepExecFailure({ scenario: err.scenario, partition: err.partition, step: err.step, error: err.initialError }));
+        const convertedResults$ = executeResults$.pipe(
+            map(result => {
+                if (isExecutionError(result)) {
+                    return AppActions.stepExecFailure({ scenario: result.scenario, partition: result.partition, step: result.step, error: result.initialError });
+                } else {
+                    const newData = convertApiDataToRiesgosData(result.state.data);
+                    return AppActions.stepExecSuccess({ scenario: result.action.scenario, partition: result.action.partition, step: result.action.step, newData });
+                }
             })
         );
 
-        return caughtAction$;
+
+        return convertedResults$;
     });
 
 
@@ -189,10 +182,14 @@ check if more  │     └───────┬──────┘
 
 class ExecutionError extends Error {
     constructor(
-        public initialError: Error, 
+        public initialError: string, 
         public scenario: ScenarioName, 
         public partition: Partition, 
         public step: string) {
-            super("Error during execution: " + initialError.message);
+            super("Error during execution: " + initialError);
         }
+}
+
+function isExecutionError(data: any): data is ExecutionError {
+    return data.initialError && data.scenario && data.partition && data.step;
 }